one less type
